+++ src/Prairie/TH.hs
@@ -11,6 +11,7 @@ import Data.Functor.Apply (Apply (..))
 import qualified Data.List as List
 import qualified Data.Text as Text
 import Data.Traversable (for)
+import GHC.Records (getField)
 import Language.Haskell.TH
 
 import Prairie.Class
@@ -126,10 +127,13 @@ mkRecord u = do
                         (compatConP (mkConstrFieldName fieldName))
                         ( NormalB $
                             VarE 'lens
-                                `AppE` VarE fieldName
+                                `AppE` (VarE 'getField `AppTypeE` LitT (StrTyLit (nameBase fieldName)))
                                 `AppE` LamE
                                     [VarP recVar, VarP newVal]
-                                    (RecUpdE (VarE recVar) [(fieldName, VarE newVal)])
+                                    ( SigE
+                                        (RecUpdE (VarE recVar) [(fieldName, VarE newVal)])
+                                        (ConT typeName)
+                                    )
                         )
                         []
         body <- CaseE (VarE arg) <$> traverse mkMatch names'types