+++ test/Prairie/DuplicateFieldSpec.hs
@@ -1,0 +1,22 @@
+{-# LANGUAGE ConstraintKinds #-}
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DuplicateRecordFields #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE GADTs #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE UndecidableInstances #-}
+
+module Prairie.DuplicateFieldSpec where
+
+import Prairie
+
+data User = User {name :: String}
+
+mkRecord ''User
+
+data Dog = Dog {name :: String}
+
+mkRecord ''Dog